feat: Trainer workout view

+++ src/services/CrudService.js
@@ -37,7 +37,33 @@ class CrudService {
    * @param {String} id - The id of the entry to get
    * @returns {Object} - An object with the data from the model
    */
+  static async getById(model, id) {
+    let entries = {};
+
+    const querySnapshot = await getDocs(collection(db, model));
+    querySnapshot.forEach((element) => {
+      if (element.id === id) {
+        const data = element.data();
+        data.id = element.id;
+        entries = { ...data };
+      }
+    });
+    return entries;
+  }
+
+  /**
+   * Get a single entry from a collection
+   * @param {String} model - The model to get the data from
+   * @param {String} id - The id of the entry to get
+   * @returns {Object} - An object with the data from the model
+   */
   static async save(model, data) {
+    if (typeof model !== 'string') {
+      throw new Error('Model must be a string');
+    }
+    if (typeof data !== 'object') {
+      throw new Error('Data must be an object');
+    }
     try {
       const resp = await addDoc(collection(db, model), {
         ...data,
@@ -58,6 +84,16 @@ class CrudService {
    * @returns {Object} - An object with the data from the model
    */
   static async update(model, id, data) {
+    if (typeof model !== 'string') {
+      throw new Error('Model must be a string');
+    }
+    if (typeof id !== 'string') {
+      throw new Error('Id must be a string');
+    }
+    if (typeof data !== 'object') {
+      throw new Error('Data must be an object');
+    }
+
     const item = doc(db, model, id);
     try {
       await updateDoc(item, { ...data, updatedAt: Timestamp.now() });
@@ -71,32 +107,18 @@ class CrudService {
   }
 
   /**
-   * Get a single entry from a collection
-   * @param {String} model - The model to get the data from
-   * @param {String} id - The id of the entry to get
-   * @returns {Object} - An object with the data from the model
-   */
-  static async getById(model, id) {
-    let entries = {};
-
-    const querySnapshot = await getDocs(collection(db, model));
-    querySnapshot.forEach((element) => {
-      if (element.id === id) {
-        const data = element.data();
-        data.id = element.id;
-        entries = { ...data };
-      }
-    });
-    return entries;
-  }
-
-  /**
    * Delete a single entry from a collection
    * @param {String} model - The model to get the data from
    * @param {String} id - The id of the entry to delete
    * @returns {Object} - An object with the data from the model
    */
   static async delete(model, id) {
+    if (typeof model !== 'string') {
+      throw new Error('Model must be a string');
+    }
+    if (typeof id !== 'string') {
+      throw new Error('Id must be a string');
+    }
     const item = doc(db, model, id);
     try {
       await deleteDoc(item);
